docker/requirements.txt
arxiv==1.4.4 ; python_version >= "3.10" and python_version < "4.0"
cachetools==5.3.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.10" and python_version < "4"
charset-normalizer==3.1.0 ; python_version >= "3.10" and python_version < "4"
deepl==1.14.0 ; python_version >= "3.10" and python_version < "4"
feedparser==6.0.10 ; python_version >= "3.10" and python_version < "4.0"
google-api-core==2.11.0 ; python_version >= "3.10" and python_version < "4.0"
google-auth==2.17.3 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-core==2.3.2 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-storage==2.8.0 ; python_version >= "3.10" and python_version < "4.0"
google-crc32c==1.5.0 ; python_version >= "3.10" and python_version < "4.0"
google-resumable-media==2.4.1 ; python_version >= "3.10" and python_version < "4.0"
googleapis-common-protos==1.59.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4"
imgkit==1.2.3 ; python_version >= "3.10" and python_version < "4.0"
nanoatp==0.3.1 ; python_version >= "3.10" and python_version < "4.0"
numpy==1.24.2 ; python_version >= "3.10" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.10" and python_version < "4.0"
pandas==2.0.0 ; python_version >= "3.10" and python_version < "4.0"
praw==7.7.0 ; python_version >= "3.10" and python_version < "4.0"
prawcore==2.3.0 ; python_version >= "3.10" and python_version < "4.0"
protobuf==4.22.3 ; python_version >= "3.10" and python_version < "4.0"
pyasn1-modules==0.2.8 ; python_version >= "3.10" and python_version < "4.0"
pyasn1==0.4.8 ; python_version >= "3.10" and python_version < "4.0"
pysbd==0.3.4 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.10" and python_version < "4.0"
requests-oauthlib==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
requests==2.28.2 ; python_version >= "3.10" and python_version < "4"
rsa==4.9 ; python_version >= "3.10" and python_version < "4"
sgmllib3k==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
slack-sdk==3.21.1 ; python_version >= "3.10" and python_version < "4.0"
tld==0.13 ; python_version >= "3.10" and python_version < "4"
tweepy==4.13.0 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.10" and python_version < "4.0"
update-checker==0.18.0 ; python_version >= "3.10" and python_version < "4.0"
urllib3==1.26.15 ; python_version >= "3.10" and python_version < "4"
websocket-client==1.5.1 ; python_version >= "3.10" and python_version < "4.0"
